+++ packages/graph-explorer/src/utils/testing/randomData.ts
@@ -424,7 +424,6 @@ export function createRandomUserStyling(): UserStyling {
 
 export function createRandomFeatureFlags(): FeatureFlags {
   return {
-    showRecoilStateLogging: createRandomBoolean(),
     showDebugActions: createRandomBoolean(),
     allowLoggingDbQuery: createRandomBoolean(),
   };